Broker based paths only index on config index
This changes the way in which broker indexes (input.broker.inputs.N) are
calculated for each input. Before it was incremented for each copy of
each item in the list. This creates a huge number of metric paths for
little value, since we usually only care about distinguishing between
different types of child input or output.

I've made the decision to therefore only prefix the metrics here by the
index of the input or output as it is configured. Meaning if you have
two inputs configured with copies set to two then the index prefixes
will now only be 0 and 1 rather than 0, 1, 2, 3.