Renew Azure service bus lock
In Azure service bus user can configure default lock time up to 5
minutes. If processing or batching takes longer then the lock is lost
and the message is retransmitted. In order to allow any processing or
batching time, this commit adds renewing this lock when it comes close
to the expiration time.