Fix ordering of resources
It's possible for input and processor resources to attempt to access
rate limits and/or cache resources. After initialisation this is fine
but there's a small gap where an eager component might attempt to access
a component before it has been initialized.

As a short term solution I've changed the ordering of construction so
that resources that typically access others are initialized last. This
won't be necessary once we can fully rule out the old style of access,
so in V4 this can be tidied up and we won't have any restrictions on
resource access.