add skip_on config to cached processor
This change adds an optional `skip_on` config field to the `cached`
processor. This field takes a bloblang mapping that must evaluate to a
boolean value. If it evaluate to `true` then the resulting batch from is
not cached. This can be useful when the result is an error and should
not be cached:

```yaml
pipeline:
  processors:
    - cached:
        key: "${! this.id }"
        # If the sql lookup fails then do not cache the error.
        skip_on: "errored()"
        processors:
          - sql_select:
              driver: "sqlite"
              dsn: "file:/sample.db"
              table: "cats"
              columns: "name,breed"
              where: "coat = 'tuxedo'"
```