Add CosmosDB components and gocosmos SQL driver
This is a revamp of #1583, where I did the initial exploration to
add these components. Since then, [gocosmos](https://pkg.go.dev/github.com/microsoft/gocosmos)
became an official Microsoft SQL driver for CosmosDB and it is
currently marked as experimental, so I decided to include it as
well because it comes with some extra goodies, such as support for
[hierarchical partition keys](https://learn.microsoft.com/en-us/azure/cosmos-db/hierarchical-partition-keys)
and [cross-partition queries](https://learn.microsoft.com/en-us/azure/cosmos-db/nosql/how-to-query-container#cross-partition-query),
which are not currently supported by the [`azcosmos`](https://pkg.go.dev/github.com/Azure/azure-sdk-for-go/sdk/data/azcosmos)
driver.

Signed-off-by: Mihai Todor <todormihai@gmail.com>