Add new processors  processor and its test in Benthos (#2334)
* Add new grouping processor and its test in Benthos

This commit introduces a new grouping processor in the Benthos package for batching multiple sub-processors together under a single processor. This new processor will allow to retain message batches and improve the readability and testability of configurations. Additionally, a new unit test has been created to ensure the processor's functionality.

Signed-off-by: Daan Gerits <daan.gerits@gmail.com>

* Refactor log message and URL in Benthos processors

The changes revamp the formatting within the Benthos package, specifically in the processor_processors.go file. The log message and URL are now consistently indented, enhancing the overall readability and maintenance later.

Signed-off-by: Daan Gerits <daan.gerits@gmail.com>

* Refactor indentation in processor_processors.go

The commit streamlines the indentation in the processor_processors.go file in the Benthos package for better readability. Subsequent maintenance work will also be facilitated due to consistent formatting for the log message and URL employed.

Signed-off-by: Daan Gerits <daan.gerits@gmail.com>

---------

Signed-off-by: Daan Gerits <daan.gerits@gmail.com>